requirements.txt
annotated-types==0.7.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
filelock==3.20.3
Flask==3.1.2
fonttools==4.61.1
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
oauthlib==3.3.1
opencv-python==4.13.0.90
packaging==26.0
pandas==3.0.0
pillow==12.1.0
playwright==1.58.0
polars==1.37.1
polars-runtime-32==1.37.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rsa==4.9.1
scipy==1.17.0
sendgrid==6.12.5
setuptools==80.10.2
six==1.17.0
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.9
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.5
